Split golf model manipulation to model.ts
