+++ src/plugins/code-golf/index.tsx
@@ -1,9 +1,9 @@
-import { Component, jsx, mountToNode } from "#DCGView";
+import { Component, jsx } from "#DCGView";
 import { AllActions } from "src/globals/extra-actions";
 import { Inserter, PluginController } from "../PluginController";
 import "./index.less";
 import { format } from "localization/i18n-core";
-import { If, IfElse, InlineMathInputView } from "src/components";
+import { If, IfElse } from "src/components";
 import {
   CalcController,
   DispatchedEvent,
@@ -11,52 +11,7 @@ import {
   FolderModel,
   ItemModel,
 } from "src/globals";
-
-function calcWidthInPixels(domNode?: HTMLElement) {
-  const rootblock = domNode?.querySelector(".dcg-mq-root-block");
-
-  if (!rootblock?.lastChild || !rootblock.firstChild) return 0;
-
-  const range = document.createRange();
-  range.setStartBefore(rootblock.firstChild);
-  range.setEndAfter(rootblock.lastChild);
-
-  const { width } = range.getBoundingClientRect();
-
-  return width;
-}
-
-function symbolCount(el: Element) {
-  const svgLen = [".dcg-mq-fraction", "svg", ".dcg-mq-token"]
-    .map((s) => el.querySelectorAll(s).length)
-    .reduce((a, b) => a + b);
-  return (
-    svgLen +
-    (el.textContent?.replace(
-      /\s|[\u00A0\u1680\u2000-\u200B\u202F\u205F\u3000]/g,
-      ""
-    )?.length ?? 0)
-  );
-}
-
-function calcSymbolCount(el?: HTMLElement) {
-  const rootblock = el?.querySelector(".dcg-mq-root-block");
-  if (!rootblock) return 0;
-
-  return rootblock ? symbolCount(rootblock) : 0;
-}
-
-interface GoodGolfStats {
-  width: number;
-  symbols: number;
-}
-
-/**
- * `TOO_LONG`: I'd love to measure this, but it's longer than `MAX_GOLF_LENGTH_CHARS`.
- *   This also blocks any future measurements of children, if this is a folder.
- * `HIDDEN`: It's offscreen, or collapsed inside a folder. No need to measure it.
- */
-export type GolfStats = "TOO_LONG" | "HIDDEN" | GoodGolfStats;
+import { populateGolfStats, GolfStats, GoodGolfStats } from "./golf-model";
 
 function _displayStats(stats: GolfStats | undefined): GoodGolfStats {
   if (!stats || stats === "TOO_LONG" || stats === "HIDDEN") {
@@ -81,47 +36,6 @@ function displayStats(stats: GolfStats | undefined) {
       elements: symbols.toString(),
     }),
   };
-}
-
-const MAX_GOLF_LENGTH_CHARS = 2000;
-
-const cachedGolfStatsPool = new Map<string, GoodGolfStats>();
-
-function getGolfStats(cc: CalcController, latex: string): GolfStats {
-  if (latex.length === 0) return { width: 0, symbols: 0 };
-
-  const cached = cachedGolfStatsPool.get(latex);
-  if (cached) return cached;
-
-  const fakeContainer = document.createElement("div");
-  document.body.appendChild(fakeContainer);
-  fakeContainer.style.transform = `scale(${1 / 0.75})`;
-
-  mountToNode(InlineMathInputView, fakeContainer, {
-    latex: () => latex ?? "",
-    isFocused: () => false,
-    selectOnFocus: () => false,
-    handleLatexChanged: () => {},
-    hasError: () => false,
-    handleFocusChanged: () => () => false,
-    ariaLabel: () => "",
-    controller: () => cc,
-  });
-
-  const stats = {
-    width: calcWidthInPixels(fakeContainer),
-    symbols: calcSymbolCount(fakeContainer),
-  };
-
-  cachedGolfStatsPool.set(latex, stats);
-
-  if (cachedGolfStatsPool.size > 10000) {
-    cachedGolfStatsPool.delete(cachedGolfStatsPool.keys().next().value!);
-  }
-
-  document.body.removeChild(fakeContainer);
-
-  return stats;
 }
 
 class GolfStatsPanel extends Component<{
@@ -222,195 +136,6 @@ export default class CodeGolf extends PluginController {
     // But that's okay, as long as it's before `updateViews` and after the
     // above changes. Desmos should be done with their `latex` mutation
     // by the end of their `handleDispatchedAction`, and ours runs after.
-    calculateMeasures(this.cc);
+    populateGolfStats(this.cc);
   }
-}
-
-/**
- * Mutate the item models, filling in `dsmGolfMeasure`.
- * This is the same approach that Desmos takes to filling in the increasing
- * sequence of index numbers from start to end. The main difference is that
- * the calculation now relies on DOM measures, which could be brutal for
- * performance. We really should do all the measures in a single pass,
- * but that is not yet implemented.
- */
-function calculateMeasures(cc: CalcController) {
-  const itemModels = cc.getAllItemModels();
-
-  // Mark long folders as too long.
-  for (let i = 0; i < itemModels.length; i++) {
-    const item = itemModels[i];
-
-    // Clear out dsmGolfStats. In particular, clearing out TOO_LONG
-    // allows re-calculate those exprs/folders if the latex gets short enough.
-    item.dsmGolfStats = undefined;
-    switch (item.type) {
-      case "expression":
-      case "folder":
-      case undefined:
-        if (isItemTooLong(itemModels, item, i)) {
-          item.dsmGolfStats = "TOO_LONG";
-        }
-        break;
-      case "image":
-      case "table":
-      case "text":
-        continue;
-      default:
-        item satisfies never;
-        continue;
-    }
-  }
-
-  // Go in reverse direction so the children's stats are computed before each folder.
-  for (let i = itemModels.length - 1; i >= 0; i--) {
-    const item = itemModels[i];
-    switch (item.type) {
-      case "expression":
-      case "folder":
-        item.dsmGolfStats = itemMeasure(cc, itemModels, item, i);
-        break;
-      case "image":
-      case "table":
-      case "text":
-      case undefined:
-        // No stats for these.
-        break;
-      default:
-        item satisfies never;
-    }
-  }
-}
-
-function isItemTooLong(
-  itemModels: ItemModel[],
-  item: ExpressionModel | FolderModel,
-  i: number
-) {
-  return item.type === "folder"
-    ? isFolderTooLong(itemModels, item, i)
-    : isExprTooLong(item);
-}
-
-function isExprTooLong(item: ExpressionModel) {
-  if (item.dsmEnableGolfDespiteLength) {
-    return false;
-  }
-  if (item.folderId) {
-    const folder = item.controller.getItemModel(item.folderId);
-    if (folder?.dsmEnableGolfDespiteLength) {
-      return false;
-    }
-  }
-  return (item.latex?.length ?? 0) > MAX_GOLF_LENGTH_CHARS;
-}
-
-/**
- * Assume folderMeasure[i] is a folder with ID `folderID`.
- * Should we just give up now for performance?
- */
-function isFolderTooLong(
-  itemModels: ItemModel[],
-  folderItem: FolderModel,
-  i: number
-): boolean {
-  if (folderItem.dsmEnableGolfDespiteLength) {
-    return false;
-  }
-  const folderId = folderItem.id;
-  let totalLength = 0;
-  for (let j = i + 1; j < itemModels.length; j++) {
-    const item = itemModels[j];
-    if (item.folderId !== folderId) break;
-    if (item.type !== "expression") continue;
-    totalLength += item.latex?.length ?? 0;
-    if (totalLength > MAX_GOLF_LENGTH_CHARS) {
-      return true;
-    }
-  }
-  return false;
-}
-
-function isHidden(item: ItemModel) {
-  // Is offscreen, hidden, or filtered out by Ctrl+F.
-  return item.renderShell || item.isHiddenFromUI || item.filteredBySearch;
-}
-
-function needToComputeMeasure(cc: CalcController, item: ItemModel): boolean {
-  if (item.dsmGolfStats === "TOO_LONG") {
-    return false;
-  } else if (!isHidden(item)) {
-    return true;
-  } else if (item.type !== "folder") {
-    const parentFolder = item.folderId && cc.getItemModel(item.folderId);
-    return !!parentFolder && needToComputeMeasure(cc, parentFolder);
-  } else {
-    return false;
-  }
-}
-
-function itemMeasure(
-  cc: CalcController,
-  itemModels: ItemModel[],
-  item: ExpressionModel | FolderModel,
-  i: number
-): GolfStats {
-  if (item.dsmGolfStats === "TOO_LONG") {
-    return "TOO_LONG";
-  }
-  if (!needToComputeMeasure(cc, item)) {
-    return "HIDDEN";
-  }
-  switch (item.type) {
-    case "expression":
-    case undefined:
-      return getGolfStats(cc, item.latex ?? "");
-    case "folder":
-      return folderMeasure(itemModels, item.id, i);
-    default:
-      item satisfies never;
-      return "HIDDEN";
-  }
-}
-
-/**
- * Assume folderMeasure[i] is a folder with ID `folderID`,
- * and the measures for everything after it have been filled in.
- * Give the GolfStats for the folder.
- */
-function folderMeasure(
-  itemModels: ItemModel[],
-  folderId: string,
-  i: number
-): GolfStats {
-  let totalWidth = 0;
-  let totalSymbols = 0;
-  for (let j = i + 1; j < itemModels.length; j++) {
-    const item = itemModels[j];
-    if (item.folderId !== folderId) break;
-    if (item.type !== "expression") continue;
-    if (
-      item.dsmGolfStats === undefined ||
-      item.dsmGolfStats === "TOO_LONG" ||
-      item.dsmGolfStats === "HIDDEN"
-    ) {
-      // To satisfy this, we need a parent folder to be TOO_LONG whenever
-      // a child expression is TOO_LONG. This is met because we use the same
-      // threshold in length for both.
-      // Also, we need children to not be HIDDEN if the parent folder is not
-      // HIDDEN. This is handled in `needToComputeMeasure`.
-
-      throw new Error(
-        `Expected dsmGolfStats to be computed for expr ${item.id}, but got ${item.dsmGolfStats}`
-      );
-    }
-    const { width, symbols } = item.dsmGolfStats;
-    totalWidth += width;
-    totalSymbols += symbols;
-  }
-
-  return {
-    width: totalWidth,
-    symbols: totalSymbols,
-  };
 }
